Task: Code improvements

+++ Task/TaskAction.struct.h
@@ -60,7 +60,7 @@ struct TaskActionEntry {
   unsigned char flags;    /* TaskAction flags. */
   datetime time_last_run; /* Time of the successful run. */
   int freq;               /* How often to run (0 for no limit). */
-  long id;                /* TaskAction's enum ID. */
+  int id;                 /* TaskAction's enum ID. */
   short tries;            /* Number of retries left (-1 for unlimited). */
   DataParamEntry args[];  /* TaskAction arguments. */
  protected:
@@ -70,7 +70,7 @@ struct TaskActionEntry {
  public:
   // Constructors.
   TaskActionEntry() : flags(0), freq(60), id(WRONG_VALUE), time_last_run(0), tries(-1) { Init(); }
-  TaskActionEntry(long _id)
+  TaskActionEntry(int _id)
       : flags(STRUCT_ENUM(TaskActionEntry, TASK_ACTION_ENTRY_FLAG_IS_ACTIVE)),
         id(_id),
         freq(60),
@@ -119,7 +119,7 @@ struct TaskActionEntry {
     SetUserError(ERR_INVALID_PARAMETER);
     return WRONG_VALUE;
   }
-  long GetId() const { return id; }
+  int GetId() const { return id; }
   // Setters.
   void TriesDec() { tries -= tries > 0 ? 1 : 0; }
   void Set(STRUCT_ENUM(TaskActionEntry, ENUM_TASK_ACTION_ENTRY_FLAG) _flag, bool _value = true) {
@@ -135,7 +135,7 @@ struct TaskActionEntry {
         freq = (int)_value;
         return;
       case TASK_ACTION_ENTRY_ID:
-        id = (long)_value;
+        id = (int)_value;
         SetFlag(STRUCT_ENUM(TaskActionEntry, TASK_ACTION_ENTRY_FLAG_IS_INVALID), id == WRONG_VALUE);
         return;
       case TASK_ACTION_ENTRY_TRIES:
